refactor(lib): moving code block to two new modules
BREAKING CHANGE: Fixed a bug inside of map sizing logic.
Auto-sizing uses the parent DOM element of the map for fitting.
However DOM padding, margin and borders were not excluded from calculations.
This produced a slightly oversized leaflet map.
If you have code to counter this effect, then make sure to remove it when upgrading.

- Adding a `operations` directory to host common renderer operations
- Moving `create-instance` to the new `operations` directory
- Moving `commit-update` to the new `operations` directory
- Moving typescript types to a new `types.ts` file
- Moving auto-sizing logic from `map.tsx` to `set-map-size.ts` file
- Updating operation did not remove nodes which are defined inside of `<MapContext.Consumer>` block
