+++ lib/index.ts
@@ -1,4 +1,4 @@
 export { default as LeafletMap } from './map'
 export * from './map'
 export { default as L } from 'leaflet'
-export { IntrinsicMapChildren, LeafletIntrinsicElements, LeafletExtensions, JSXRenderer } from './types'
+export { IntrinsicMapChildren, LeafletIntrinsicElements, LeafletExtensions, JSXRenderer } from './catalog'
