+++ lib/renderer.ts
@@ -1,340 +1,13 @@
-/// <reference types="../@types" />
-
-import ReactReconciler, { HostConfig } from 'react-reconciler'
+import ReactReconciler from 'react-reconciler'
 import L from 'leaflet'
-import { isEqualWith, isFunction } from 'lodash'
+import { isEqualWith, isFunction, noop } from 'lodash'
 
-import Marker from 'leaflet/dist/images/marker-icon.png'
-import MarkerShadow from 'leaflet/dist/images/marker-shadow.png'
+import { JSXRenderer, LeafletIntrinsicElements, LeafletExtensions } from './catalog'
+import { ElementProps, RenderHostConfigs } from './types'
+import { Add, Remove } from './cache'
 
-import { JSXRenderer, LeafletIntrinsicElements, LeafletExtensions } from './types'
-
-type Type = keyof LeafletIntrinsicElements
-type Props = LeafletIntrinsicElements[Type]
-type Container = HTMLDivElement
-type HostContext = Record<string, unknown>
-type UpdatePayload = {
-  rootContainer: Container
-  hostContext: HostContext
-}
-type Instance = {
-  type: Type
-  category: 'map' | 'layergroup' | 'featuregroup' | 'layer' | 'control' | 'handler'
-  leaflet: L.Class
-  props: unknown
-  parent?: Instance
-} | undefined
-type RenderHostConfigs = HostConfig<Type, Props, Container, Instance, Instance, Instance, Instance, L.Map, HostContext, UpdatePayload, any, any, any>
-
-const instances = new Map<HTMLElement, ReactReconciler.FiberRoot>()
-const maps = new Map<HTMLElement, L.Map>()
-let contentRenderer: JSXRenderer | undefined
-
-/**
- * This method should return a newly created node. For example, the DOM renderer would call `document.createElement(type)` here and then set the properties from `props`.
- *
- * You can use `rootContainer` to access the root container associated with that tree. For example, in the DOM renderer, this is useful to get the correct `document` reference that the root belongs to.
- *
- * The `hostContext` parameter lets you keep track of some information about your current place in the tree. To learn more about it, see `getChildHostContext` below.
- *
- * The `internalHandle` data structure is meant to be opaque. If you bend the rules and rely on its internal fields, be aware that it may change significantly between versions. You're taking on additional maintenance risk by reading from it, and giving up all guarantees if you write something to it.
- *
- * This method happens **in the render phase**. It can (and usually should) mutate the node it has just created before returning it, but it must not modify any other nodes. It must not register any event handlers on the parent tree. This is because an instance being created doesn't guarantee it would be placed in the tree — it could be left unused and later collected by GC. If you need to do something when an instance is definitely in the tree, look at `commitMount` instead.
- */
-const createInstance = (type: Type, props: Props, rootContainer: Container, hostContext: any, internalHandle: any) => {
-  let instance: Instance
-
-  switch (type) {
-    case 'lfMap': {
-      const { options, whenReady, ...restProps } = props as LeafletIntrinsicElements['lfMap']
-      const mp = L.map(rootContainer, options)
-      if (whenReady) {
-        mp.whenReady(() => whenReady(mp))
-      }
-      maps.set(rootContainer, mp)
-
-      instance = {
-        type,
-        category: 'map',
-        leaflet: mp,
-        props
-      }
-      break
-    }
-
-    case 'lfImage': {
-      const lfImageProps = props as LeafletIntrinsicElements['lfImage']
-      const layer = L.imageOverlay(lfImageProps.imageUrl, lfImageProps.bounds, lfImageProps.options)
-      setProps(layer, lfImageProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfPopup': {
-      const { children, latlng, ...restProps } = props as LeafletIntrinsicElements['lfPopup']
-      const layer = L.popup(restProps.options)
-      const element = document.createElement('section') // TODO: imrove the wrapper element for lfPopup
-      layer.setContent(element)
-      setProps(layer, restProps)
-      if (contentRenderer) {
-        contentRenderer(children as any, element)
-      }
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfTooltip': {
-      const { children, ...restProps } = props as LeafletIntrinsicElements['lfTooltip']
-      const layer = L.tooltip(restProps.options)
-      const element = document.createElement('section') // TODO: imrove the wrapper element for lfPopup
-      layer.setContent(element)
-      setProps(layer, restProps)
-      if (contentRenderer) {
-        contentRenderer(children as any, element)
-      }
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfRectangle': {
-      const { bounds, options, ...restProps } = props as LeafletIntrinsicElements['lfRectangle']
-      const layer = L.rectangle(bounds, options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfMarker': {
-      const { latlng, ...restProps } = props as LeafletIntrinsicElements['lfMarker']
-      const icon = L.icon({
-        iconUrl: Marker,
-        shadowUrl: MarkerShadow
-      })
-      const layer = L.marker(latlng, { icon, ...restProps.options })
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfTiles': {
-      const { urlTemplate, ...restProps } = props as LeafletIntrinsicElements['lfTiles']
-      const layer = L.tileLayer(urlTemplate, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfTilesWMS': {
-      const { baseUrl, ...restProps } = props as LeafletIntrinsicElements['lfTilesWMS']
-      const layer = L.tileLayer.wms(baseUrl, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfVideo': {
-      const { video, bounds, ...restProps } = props as LeafletIntrinsicElements['lfVideo']
-      const layer = L.videoOverlay(video, bounds, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfPolyline': {
-      const { latlngs, ...restProps } = props as LeafletIntrinsicElements['lfPolyline']
-      const layer = L.polyline(latlngs as any, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfPolygon': {
-      const { latlngs, ...restProps } = props as LeafletIntrinsicElements['lfPolygon']
-      const layer = L.polygon(latlngs, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfCircle': {
-      const { latlng, ...restProps } = props as LeafletIntrinsicElements['lfCircle']
-      const layer = L.circle(latlng, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfCircleMarker': {
-      const { latlng, ...restProps } = props as LeafletIntrinsicElements['lfCircleMarker']
-      const layer = L.circleMarker(latlng, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfSVG': {
-      const { svgImage, bounds, ...restProps } = props as LeafletIntrinsicElements['lfSVG']
-      const layer = L.svgOverlay(svgImage, bounds, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfLayerGroup': {
-      const { children, ...restProps } = props as LeafletIntrinsicElements['lfLayerGroup']
-      const layer = L.layerGroup([], restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layergroup',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfFeatureGroup': {
-      const { children, ...restProps } = props as LeafletIntrinsicElements['lfFeatureGroup']
-      const layer = L.featureGroup([], restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'featuregroup',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfGeoJSON': {
-      const { geojson, ...restProps } = props as LeafletIntrinsicElements['lfGeoJSON']
-      const layer = L.geoJSON(geojson, restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-
-    case 'lfGridLayer': {
-      const restProps = props as LeafletIntrinsicElements['lfGridLayer']
-      const layer = L.gridLayer(restProps.options)
-      setProps(layer, restProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: layer,
-        props
-      }
-      break
-    }
-  }
-
-  if (!instance) {
-    if (type.indexOf('Layer') > -1) {
-      const layerProps = props as LeafletExtensions.Layer
-      const klassInstance = new layerProps.klass({ ...layerProps.params, children: props.children, jsxRenderer: contentRenderer })
-      setProps(klassInstance, layerProps)
-      instance = {
-        type,
-        category: 'layer',
-        leaflet: klassInstance,
-        props
-      }
-    } else if (type.indexOf('Control') > -1) {
-      const controlProps = props as LeafletExtensions.Control
-      const klassInstance = new controlProps.klass({ ...controlProps.params, children: props.children, jsxRenderer: contentRenderer })
-      instance = {
-        type,
-        category: 'control',
-        leaflet: klassInstance,
-        props
-      }
-    } else if (type.indexOf('Handler') > -1) {
-      const handlerProps = props as LeafletExtensions.Handler
-      instance = {
-        type,
-        category: 'handler',
-        leaflet: handlerProps.klass,
-        props
-      }
-    }
-
-    if (type.startsWith('lf') && !instance) {
-      throw Error(`leaflet-react-fibers: Unknown type ${type}. If you are trying to use customized JSX, then make sure that your intrinsic declaration (aka tag name) ends with one of "Control", "Layer" or "Handler".`)
-    }
-  }
-
-  return instance
-}
+import createInstance from './operations/create-instance'
+import commitUpdate from './operations/commit-update'
 
 const hostConfig: RenderHostConfigs = {
   supportsMutation: true,
@@ -358,6 +31,7 @@ const hostConfig: RenderHostConfigs = {
   appendInitialChild: (parentInstance, childInstance) => {
     if (!parentInstance || !childInstance) return
 
+    // Sync parent instance for future reference
     childInstance.parent = parentInstance
 
     if (parentInstance.category === 'map' && childInstance.category === 'handler') {
@@ -533,21 +207,23 @@ const hostConfig: RenderHostConfigs = {
   *
   * Although this method currently runs in the commit phase, you still should not mutate any other nodes in it. If you need to do some additional work when a node is definitely connected to the visible tree, look at `commitMount`.
   */
-  appendChild: (parentInstance, child) => {
-    if (!parentInstance || !child) return
+  appendChild: (parentInstance, childInstance) => {
+    if (!childInstance) return
 
-    child.parent = parentInstance
+    // Sync parent for future reference
+    childInstance.parent = parentInstance
 
     switch (parentInstance?.category) {
       case 'map':
       case 'layergroup':
-      case 'featuregroup':
+      case 'featuregroup': {
         const parent = parentInstance.leaflet as L.Map | L.LayerGroup
-        const layer = child?.leaflet as L.Layer
+        const layer = childInstance.leaflet as L.Layer
 
         parent.addLayer(layer)
 
         break
+      }
     }
   },
 
@@ -564,13 +240,18 @@ const hostConfig: RenderHostConfigs = {
   *
   * Note that React uses this method both for insertions and for reordering nodes. Similar to DOM, it is expected that you can call `insertBefore` to reposition an existing child. Do not mutate any other parts of the tree from it.
   */
-  insertBefore: (parentInstance, child, beforeChild) => {
+  insertBefore: (parentInstance, childInstance, beforeChild) => {
+    if (!childInstance) return
+
+    // Sync parent for future reference
+    childInstance.parent = parentInstance
+
     switch (parentInstance?.category) {
       case 'map':
       case 'layergroup':
-      case 'featuregroup':
+      case 'featuregroup': {
         const parent = parentInstance.leaflet as L.Map | L.LayerGroup
-        const layer = child?.leaflet as L.Layer
+        const layer = childInstance.leaflet as L.Layer
 
         parent.addLayer(layer)
 
@@ -578,7 +259,7 @@ const hostConfig: RenderHostConfigs = {
         // Sometimes they do not since leaflet layer insertions do not match JSX defs.
         // E.g. all lfCircle are added to a common <svg> element so if we place one next to a lfTileLayer,
         // then they will be added to different DOM elements and insertBefore is meaningless
-        const childAny = child?.leaflet as any
+        const childAny = childInstance?.leaflet as any
         const beforeChildAny = beforeChild?.leaflet as any
         const element = childAny.getElement ? childAny.getElement() : childAny.getContainer ? childAny.getContainer : null
         const elementBefore = beforeChildAny.getElement ? beforeChildAny.getElement() : beforeChildAny.getContainer ? beforeChildAny.getContainer : null
@@ -589,13 +270,9 @@ const hostConfig: RenderHostConfigs = {
         }
 
         break
+      }
     }
   },
-
-  /**
-  * Same as `insertBefore`, but for when a node is attached to the root container. This is useful if attaching to the root has a slightly different implementation, or if the root container nodes are of a different type than the rest of the tree.
-  */
-  // insertInContainerBefore: (container, child, beforeChild) => { },
 
   /**
   * This method should mutate the `parentInstance` to remove the `child` from the list of its children.
@@ -606,11 +283,13 @@ const hostConfig: RenderHostConfigs = {
     switch (parentInstance?.category) {
       case 'map':
       case 'layergroup':
-      case 'featuregroup':
+      case 'featuregroup': {
         const parent = parentInstance.leaflet as L.Map | L.LayerGroup
         const layer = child?.leaflet as L.Layer
         parent.removeLayer(layer)
+
         break
+      }
     }
   },
 
@@ -619,22 +298,9 @@ const hostConfig: RenderHostConfigs = {
   */
   removeChildFromContainer: (container, child) => {
     const map = child?.leaflet as L.Map
+    Remove(container)
     map?.remove()
   },
-
-  /**
-  * If you returned `true` from `shouldSetTextContent` for the previous props, but returned `false` from `shouldSetTextContent` for the next props, React will call this method so that you can clear the text content you were managing manually. For example, in the DOM you could set `node.textContent = ''`.
-  *
-  * If you never return `true` from `shouldSetTextContent`, you can leave it empty.
-  */
-  // resetTextContent: (instance) => { },
-
-  /**
-   * This method should mutate the `textInstance` and update its text content to `nextText`.
-   *
-   * Here, `textInstance` is a node created by `createTextInstance`.
-   */
-  // commitTextUpdate: (textInstance, oldText, newText) => { },
 
   /**
   * This method is only called if you returned `true` from `finalizeInitialChildren` for this instance.
@@ -714,127 +380,7 @@ const hostConfig: RenderHostConfigs = {
   *
   * The `internalHandle` data structure is meant to be opaque. If you bend the rules and rely on its internal fields, be aware that it may change significantly between versions. You're taking on additional maintenance risk by reading from it, and giving up all guarantees if you write something to it.
   */
-  commitUpdate: (instance, updatePayload, type, prevProps, nextProps, internalHandle) => {
-    const map = maps.get(updatePayload.rootContainer)
-
-    if (!map) {
-      throw Error(`leaflet-react-fibers: Did not find a map while updating instance of type "${type}" and category of "${instance?.category}"`)
-    }
-
-    switch (instance?.category) {
-      case 'map': {
-        map.options = { ...prevProps.options as L.MapOptions, ...nextProps.options as L.MapOptions }
-        map.invalidateSize({ animate: false })
-        break
-      }
-
-      case 'handler': {
-        const parent = map as L.Map & { [key: string]: L.Handler }
-        const props = nextProps as LeafletExtensions.Handler
-
-        if (props.enabled) {
-          parent[props.name].enable()
-        } else {
-          parent[props.name].disable()
-        }
-
-        break
-      }
-
-      case 'layer':
-      case 'layergroup':
-      case 'featuregroup': {
-        const layer = instance?.leaflet as L.Layer & Partial<LeafletExtensions.Stateful<any>>
-
-        // special case of attached popup
-        if (instance.type === 'lfPopup') {
-          const popup = instance.leaflet as L.Popup
-          const popupProps = nextProps as LeafletIntrinsicElements['lfPopup']
-
-          if (popupProps.latlng) {
-            popup.setLatLng(popupProps.latlng)
-          }
-
-          if (instance.parent?.category === 'map') {
-            const parent = instance.parent?.leaflet as L.Map
-
-            if (popupProps.isOpen) {
-              parent.openPopup(popup)
-            } else {
-              parent.closePopup(popup)
-            }
-          } else {
-            const parent = instance.parent?.leaflet as L.Layer
-
-            if (popupProps.isOpen) {
-              parent.openPopup()
-            } else {
-              parent.closePopup()
-            }
-          }
-        } else {
-          if (instance.parent?.category === 'map' || instance.parent?.category === 'featuregroup' || instance.parent?.category === 'layergroup') {
-            const parent = instance.parent?.leaflet as L.Map | L.FeatureGroup | L.LayerGroup
-            const newInstance = createInstance(type, nextProps, updatePayload.rootContainer, updatePayload.hostContext, internalHandle)
-
-            if (newInstance) {
-              const newLayer = newInstance.leaflet as L.Layer & Partial<LeafletExtensions.Stateful<any>>
-              const layerProps = nextProps as LeafletExtensions.Layer
-
-              // stateful layer, default behavior
-              if (!layerProps.hasOwnProperty('mutable') || layerProps.mutable) {
-                if (layer.getState && newLayer.setState) {
-                  const oldState = layer.getState()
-                  newLayer.setState(oldState)
-                }
-
-                parent.removeLayer(layer)
-                parent.addLayer(newLayer)
-              }
-              // standard layer
-              else {
-                // not much can change after layer is created, note that popups and tooltips are handled via JSX directly
-              }
-
-              instance.leaflet = newInstance.leaflet
-            }
-          }
-        }
-
-        break
-      }
-
-      case 'control': {
-        const parent = instance.parent?.leaflet as L.Map
-        const control = instance?.leaflet as L.Control & Partial<LeafletExtensions.Stateful<any>>
-        const controlProps = nextProps as LeafletExtensions.Control['params']
-        const newInstance = createInstance(type, nextProps, updatePayload.rootContainer, updatePayload.hostContext, internalHandle)
-
-        if (newInstance) {
-          const newControl = newInstance.leaflet as L.Control & Partial<LeafletExtensions.Stateful<any>>
-
-          // stateful control, default behavior
-          if (!controlProps.hasOwnProperty('mutable') || controlProps.mutable) {
-            if (control.getState && newControl.setState) {
-              const oldState = control.getState()
-              newControl.setState(oldState)
-            }
-
-            parent.removeControl(control)
-            parent.addControl(newControl)
-          }
-          // standard control
-          else {
-            control.setPosition(controlProps.position)
-          }
-
-          instance.leaflet = newInstance.leaflet
-        }
-
-        break
-      }
-    }
-  },
+  commitUpdate,
 
   /**
   * This method should make the `instance` invisible without removing it from the tree. For example, it can apply visual styling to hide it. It is used by Suspense to hide the tree while the fallback is visible.
@@ -872,79 +418,9 @@ const hostConfig: RenderHostConfigs = {
   // -------------------
 
   supportsHydration: false
-
-  // canHydrateInstance: (instance, type, props) => { },
-
-  // canHydrateTextInstance: (instance, text) => { },
-
-  // canHydrateSuspenseInstance: (instance) => { },
-
-  // isSuspenseInstancePending: (instance) => { },
-
-  // isSuspenseInstanceFallback: (instance) => { },
-
-  // registerSuspenseInstanceRetry: (instance, callback) => { },
-
-  // getNextHydratableSibling: (instance) => { },
-
-  // getFirstHydratableChild: (parentInstance) => { },
-
-  // hydrateInstance: (instance, type, props, rootContainerInstance, hostContext, internalInstanceHandle) => { },
-
-  // hydrateTextInstance: (textInstance, text, internalInstanceHandle) => { },
-
-  // hydrateSuspenseInstance: (suspenseInstance, internalInstanceHandle) => { },
-
-  // getNextHydratableInstanceAfterSuspenseInstance: (suspenseInstance) => { },
-
-  // // Returns the SuspenseInstance if this node is a direct child of a
-  // // SuspenseInstance. I.e. if its previous sibling is a Comment with
-  // // SUSPENSE_x_START_DATA. Otherwise, null.
-  // getParentSuspenseInstance: (targetInstance) => { },
-
-  // commitHydratedContainer: (container) => { },
-
-  // commitHydratedSuspenseInstance: (suspenseInstance) => { },
-
-  // didNotMatchHydratedContainerTextInstance: (parentContainer, textInstance, text) => { },
-
-  // didNotMatchHydratedTextInstance: (parentType, parentProps, parentInstance, textInstance, text) => { },
-
-  // didNotHydrateContainerInstance: (parentContainer, instance) => { },
-
-  // didNotHydrateInstance: (parentType, parentProps, parentInstance, instance) => { },
-
-  // didNotFindHydratableContainerInstance: (parentContainer, type, props) => { },
-
-  // didNotFindHydratableContainerTextInstance: (parentContainer, text) => { },
-
-  // didNotFindHydratableContainerSuspenseInstance: (parentContainer) => { },
-
-  // didNotFindHydratableInstance: (parentType, parentProps, parentInstance, type, props) => { },
-
-  // didNotFindHydratableTextInstance: (parentType, parentProps, parentInstance, text) => { },
-
-  // didNotFindHydratableSuspenseInstance: (parentType, parentProps, parentInstance) => { },
-
 }
 
-const setProps = function <P extends { [key: string]: any }>(leaflet: L.Class & { [key: string]: any }, props: P) {
-  const keys = Object.keys(props)
-  keys.forEach((key) => {
-    if (isFunction(props[key]) && leaflet.on) {
-      const leafletKey = key.startsWith('on') ? key.substring(2).toLocaleLowerCase('en-US') : key
-      leaflet.on(leafletKey, props[key])
-    }
-
-    const setter = 'set' + key[0].toLocaleUpperCase('en-US') + (key.length > 1 ? key.substr(1) : '')
-
-    if (leaflet[setter] && isFunction(leaflet[setter])) {
-      leaflet[setter](props[key])
-    }
-  })
-}
-
-const propsChanged = (oldProps: Props, nextProps: Props) => {
+const propsChanged = (oldProps: ElementProps, nextProps: ElementProps) => {
   const isEqual = isEqualWith(oldProps, nextProps, (a: any, b: any) => {
     // non-memoized functions should not result in layer re-renders, compare function bodies instead
     if (isFunction(a) || isFunction(b)) {
@@ -954,15 +430,18 @@ const propsChanged = (oldProps: Props, nextProps: Props) => {
   return !isEqual
 }
 
+const instances = new Map<HTMLElement, ReactReconciler.FiberRoot>()
 const renderer = ReactReconciler(hostConfig)
+
 export default {
-  render: (reactElement: React.ReactNode, domElement: HTMLDivElement, jsxRenderer?: JSXRenderer, callback?: () => void) => {
-    contentRenderer = jsxRenderer
+  render: (reactElement: globalThis.React.ReactNode, domElement: HTMLDivElement, jsxRenderer?: JSXRenderer, callback?: () => void) => {
+    Add('jsxRenderer', jsxRenderer)
+
     let container
     if (instances.has(domElement)) {
       container = instances.get(domElement)
     } else {
-      container = renderer.createContainer(domElement, 1, false, { onDeleted: callback, onHydrated: callback })
+      container = renderer.createContainer(domElement, 1, null, false, null, 'leafletReactFibers', noop, null)
       instances.set(domElement, container)
     }
